+++ policyengine_us/tools/taxsim/generate_taxsim_tests.py
@@ -104,7 +104,6 @@ class TaxSim35:
     NEUTRALIZED_VARIABLES = [
         "snap",
         "ssi",
-        "state_supplement",
         "tanf",
         "wic",
     ]